git-svn-id: svn://123.57.173.36/meijia-parent@1519 52ac4245-7803-4e87-b018-04b7c5ce731e

+++ simi-dao/src/main/java/com/simi/po/dao/op/AppToolsMapper.java
@@ -9,7 +9,7 @@ public interface AppToolsMapper {
 
     int insert(AppTools record);
 
-    int insertSelective(AppTools record);
+    Long insertSelective(AppTools record);
 
     AppTools selectByPrimaryKey(Long tId);
